Split game state schema into multiple files
