--- src/utils/schemas/gameState.json
@@ -1,51 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "definitions": {
-    "ExampleRoom1": {
-      "type": "object",
-      "properties": {
-        "type": {
-          "const": "exampleRoom1"
-        },
-        "playerState": {
-          "enum": [
-            "AtEntrance"
-          ]
-        }
-      },
-      "required": ["type", "playerState"],
-      "additionalProperties": false
-    },
-    "ExampleRoom2": {
-      "type": "object",
-      "properties": {
-        "type": {
-          "const": "exampleRoom2"
-        },
-        "playerState": {
-          "enum": [
-            "AtEntrance",
-            "State2A",
-            "State2B"
-          ]
-        }
-      },
-      "required": ["type", "playerState"],
-      "additionalProperties": false
-    },
-    "RoomState": {
-      "type": "object",
-      "oneOf": [
-        { "$ref": "#/definitions/ExampleRoom1" },
-        { "$ref": "#/definitions/ExampleRoom2" }
-      ]
-    }
-  },
-  "title": "GameState",
-  "type": "object",
-  "properties": {
-    "roomState": { "$ref": "#/definitions/RoomState" }
-  },
-  "required": ["roomState"],
-  "additionalProperties": false
-}
