+++ src/utils/schemas/normalized/exampleRoom1.json
@@ -1,0 +1,17 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "title": "ExampleRoom1",
+  "type": "object",
+  "properties": {
+    "type": {
+      "const": "exampleRoom1"
+    },
+    "playerState": {
+      "enum": [
+        "AtEntrance"
+      ]
+    }
+  },
+  "required": ["type", "playerState"],
+  "additionalProperties": false
+}
