+++ src/utils/schemas/normalized/gameState.json
@@ -1,0 +1,10 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "title": "GameState",
+  "type": "object",
+  "properties": {
+    "roomState": { "$ref": "./roomState.json" }
+  },
+  "required": ["roomState"],
+  "additionalProperties": false
+}
