+++ src/utils/schemas/normalized/roomState.json
@@ -1,0 +1,9 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "title": "RoomState",
+  "type": "object",
+  "oneOf": [
+    { "$ref": "./exampleRoom1.json" },
+    { "$ref": "./exampleRoom2.json" }
+  ]
+}
