+++ tests/testHelpers/generateGameState.ts
@@ -1,10 +1,9 @@
 import { schemaToGenerator } from '@randograms/schema-to-generator';
-import { gameStateSchema } from '../../src/utils/schemas/gameStateSchema';
-import { roomStateSchema } from '../../src/utils/schemas/roomStateSchema';
+import gameStateSchema from '../../src/utils/schemas/normalized/gameState.json';
+import roomStateSchema from '../../src/utils/schemas/normalized/roomState.json';
 import { GameState, RoomState } from '../../src/utils/types';
 
-export const generateGameState = schemaToGenerator<GameState>(
-  gameStateSchema as Record<string, unknown>,
-);
+// TODO: schemaToGenerator cannot handle $refs, so this needs a denormalized schema
+export const generateGameState = schemaToGenerator<GameState>(gameStateSchema);
 
 export const generateRoomState = schemaToGenerator<RoomState>(roomStateSchema);
